working on  clients

+++ database/migrations/2024_02_13_212054_create_clients_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('tva');
             $table->string('company');
             $table->string('adresse');
-            $table->smallInteger('code_postal');
+            $table->integer('code_postal');
             $table->string('ville');
             $table->string('email');
         });
